feat: add user admin and authentication; finish part 4

+++ part4/controllers/blogs.js
@@ -2,15 +2,34 @@ const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
 
 blogsRouter.get('/', async (request, response) => {
-  const blog = await Blog.find({});
+  const blog = await Blog.find({}).populate('user', { username: 1, name: 1 });
+  response.json(blog);
+});
+
+blogsRouter.get('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id).populate('user', { username: 1, name: 1 });
   response.json(blog);
 });
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body);
+  const body = request.body;
+  const user = request.user;
 
-  const result = await blog.save();
-  response.status(201).json(result);
+  if (user === null) return response.status(400).json({ error: 'userId missing or not valid' });
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id
+  });
+
+  const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
+  response.status(201).json(savedBlog);
 });
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -29,8 +48,15 @@ blogsRouter.put('/:id', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  const blog = await Blog.findById(request.params.id);
+  if (request.user === null) return response.status(400).json({ error: 'userId missing or not valid' });
+
+  if ( blog.user.toString() === request.user._id.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  } else {
+    response.status(401).json({ error: 'token invalid' });
+  }
 });
 
 module.exports = blogsRouter;
